# model/entity/article.gen.go
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.

package entity

import (
	"time"

	"gorm.io/gorm"
)

const TableNameArticle = "article"

// Article mapped from table <article>
type Article struct {
	ID        uint64         `gorm:"column:id;type:bigint unsigned;primaryKey;autoIncrement:true" json:"id"`
	CreatedAt time.Time      `gorm:"column:created_at;type:datetime(3)" json:"created_at"`
	UpdatedAt time.Time      `gorm:"column:updated_at;type:datetime(3)" json:"updated_at"`
	DeletedAt gorm.DeletedAt `gorm:"column:deleted_at;type:datetime(3);index:idx_article_deleted_at,priority:1" json:"deleted_at"`
	Title     string         `gorm:"column:title;type:varchar(64);not null" json:"title"`                              // 标题
	Img       string         `gorm:"column:img;type:varchar(128)" json:"img"`                                          // 首图
	Desc      string         `gorm:"column:desc;type:varchar(255)" json:"desc"`                                        // 描述
	Content   string         `gorm:"column:content;type:longtext" json:"content"`                                      // 内容
	Cid       uint64         `gorm:"column:cid;type:bigint unsigned;index:fk_category_articles,priority:1" json:"cid"` // 分类ID
}

// TableName Article's table name
func (*Article) TableName() string {
	return TableNameArticle
}